install-local.sh: use type=NodePort for ingress-nginx
When we are installing ingress-nginx, set it to use a ClusterIP
for the service instead of a LoadBalancer since kind doesn't
install a LoadBalancer by default. Also set hostPort so that the
the ingress is expoed outside of kind (this is in combination
with the existing extraPortMappings configured in
install-local/install-local-kind-config.yaml).

See: https://github.com/OpenNMS/helm-charts/commit/c27776f5573227f8b4f1d3393720e3475846dcf4
And: https://github.com/kubernetes-sigs/kind/issues/2889#issuecomment-1321223613