Run tools/ssl-check.sh in Tilt
Added a "-t <tries>" option to ssl-check.sh to give it a few
attempts to verify an HTTPS connection to the UI ingress because it
usually takes a short time for ingress-nginx to switch to the new
certificate after it is reloaded.

Also made ssl-check.sh require the domain and port be passed on the
command-line.