ssl-check.sh: remove effective no-op of '2>&1' before '> /dev/null'
The shell has somewhat odd handling of redirection that causes
'2>&1 > /dev/null' to redirect stderr to the terminal and then redirect
stdout to /dev/null. This will leave stderr redirected to stdout.
To redirect both to /dev/null, the '> /dev/null' redirect must come
first, like this: '> /dev/null 2>&1'.

Or, with bash, this can be used: '&>/dev/null'.

But most of the checks have been quieted down by default, so I want
to see errors now, so we just remove the '2>&1' now.