Add OpenTelemetry Java agent to the Minion
The Java agent is added to the container, but it is currently only
enabled by default in the lokahi Helm chart, so pretty much only
for testing with Tilt and install-local.

Need to eventually migrate the existing in-app opentracing/Jaeger
tracing to OpenTelemetry, but right now, this is useful for
tracking low-level gRPC connection events (but not individual
messages in streaming connections).