Helm chart: allow setting the image version in one place
Also had to tweak how the existing "deployment.env" define is
included since it how has to call the new "lokahi.image" define.

You can only pass one set of values with include, but I found this
great hack where you can build a custom dictionary to both give
access to global values, but also local values:
https://stackoverflow.com/a/52648558

We use this to pass a reference to the top of the Values tree,
and also to a service-specific part of the tree (for example:
Values.OpenNMS.Alert).