LOK-1983: log flow received and completion metrics near each other
Record the addTenantFlowReceviedCount metrics as close as possible
to the addTenantFlowCompletedCount (i.e.: after processing) so both
hopefully end up together in results so we can compute a reasonable
success rate. When processing takes a long time (sometimes > 1 second
due to LOK-1981), recording the received and completion counts with
a delay leads to an odd success rate chart that isn't very useful
when it periodically varies from 80 - 140% due to recording timing.