DP-90: Update Locations tour
DP-90: Update Locations tour - as part of the work to update the Locations tour in Pendo, I noticed changes in the UI and procedure for deploying a Minion updated in the UI, but not the docs. Updated the docs to reflect the new reality.

Also, removed references to the "Quick Action" button that is no longer on the Insights Dashboard. Made sure the docs describe the Top N and Node reachability charts as well.