release.sh: don't update develop
This change stops trying to create CHANGELOG files in develop and just
creates them directly in release. I'll probably go back and remove the
CHANGELOG files in develop so there isn't confusion about why they
abruptly stopped there.

Previously, we got these errors when we tried to push to develop and it was
denied because we have required status checks for that branch:

    remote: error: GH006: Protected branch update failed for refs/heads/develop.
    remote: error: 18 of 18 required status checks are expected.

-- https://github.com/OpenNMS-Cloud/lokahi/actions/runs/7049125718/job/19186877109#step:3:425